+++ trunk/src/dataWins.h
@@ -8,7 +8,7 @@ class dataWins {
 	public:
 			dataWins();
 			~dataWins();
-			dataWins(unsigned short int, unsigned long int, unsigned long int,int,int,double,double,double);
+			dataWins(unsigned short int, unsigned long int, unsigned long int,int,double,double,double,double);
 			dataWins(const dataWins&);
 			bool operator <(dataWins) const;
 
@@ -16,7 +16,7 @@ class dataWins {
 			unsigned long int start;//8
 			unsigned long int stop;
 			int eCount;
-			int iCount;
+			double iCount;
 			double gcPerc;
 			double alignPerc;
 			double cnvScore;
